Move AspNetCore tests to their own namespace
