+++ tests/Agent/IntegrationTests/IntegrationTests/AspNetCore/AspNetCoreMvcBasicRequestsTests.cs
@@ -11,16 +11,16 @@ using NewRelic.Testing.Assertions;
 using Xunit;
 using Xunit.Abstractions;
 
-namespace NewRelic.Agent.IntegrationTests
+namespace NewRelic.Agent.IntegrationTests.AspNetCore
 {
     [NetCoreTest]
-    public class AspNetCoreGenericWebHostTests : IClassFixture<RemoteServiceFixtures.AspNetCore3FeaturesFixture>
+    public class AspNetCoreMvcBasicRequestsTests : IClassFixture<RemoteServiceFixtures.AspNetCoreMvcBasicRequestsFixture>
     {
-        private readonly RemoteServiceFixtures.AspNetCore3FeaturesFixture _fixture;
+        private readonly RemoteServiceFixtures.AspNetCoreMvcBasicRequestsFixture _fixture;
 
         private const int ExpectedTransactionCount = 2;
 
-        public AspNetCoreGenericWebHostTests(RemoteServiceFixtures.AspNetCore3FeaturesFixture fixture, ITestOutputHelper output)
+        public AspNetCoreMvcBasicRequestsTests(RemoteServiceFixtures.AspNetCoreMvcBasicRequestsFixture fixture, ITestOutputHelper output)
         {
             _fixture = fixture;
             _fixture.TestLogger = output;
@@ -34,6 +34,7 @@ namespace NewRelic.Agent.IntegrationTests
                 },
                 exerciseApplication: () =>
                 {
+                    _fixture.GetCORSPreflight();
                     _fixture.Get();
                     _fixture.ThrowException();
 
@@ -49,10 +50,13 @@ namespace NewRelic.Agent.IntegrationTests
             var metrics = _fixture.AgentLog.GetMetrics().ToList();
             Assert.NotNull(metrics);
 
+            Assertions.MetricsExist(_generalMetrics, metrics);
+            Assertions.MetricsDoNotExist(_unexpectedMetrics, metrics);
+
             NrAssert.Multiple
             (
-                () => Assertions.MetricsExist(_generalMetrics, metrics),
-                () => Assertions.MetricsExist(_indexMetrics, metrics)
+                () => Assertions.MetricsExist(_indexMetrics, metrics),
+                () => Assertions.MetricsExist(_throwMetrics, metrics)
             );
 
             var expectedTransactionTraceSegments = new List<string>
@@ -64,8 +68,10 @@ namespace NewRelic.Agent.IntegrationTests
             var transactionSample = _fixture.AgentLog.GetTransactionSamples()
                 .FirstOrDefault(sample => sample.Path == @"WebTransaction/MVC/Home/Index");
 
+
             Assert.NotNull(transactionSample);
             Assertions.TransactionTraceSegmentsExist(expectedTransactionTraceSegments, transactionSample);
+
 
             var expectedErrorEventAttributes = new Dictionary<string, string>
             {
@@ -125,5 +131,12 @@ namespace NewRelic.Agent.IntegrationTests
             new Assertions.ExpectedMetric { metricName = @"DotNet/HomeController/ThrowException", metricScope = @"WebTransaction/MVC/Home/ThrowException", callCount = 1 },
             new Assertions.ExpectedMetric { metricName = @"DotNet/Middleware Pipeline", metricScope = @"WebTransaction/MVC/Home/ThrowException", callCount = 1 }
         };
+
+        private readonly List<Assertions.ExpectedMetric> _unexpectedMetrics = new List<Assertions.ExpectedMetric>
+        {
+			// SUPNET-492 (ASP.NET Core CORS MGI)
+			new Assertions.ExpectedMetric {metricName = @"WebTransaction/ASP/Home/About", callCount = 1},
+        };
+
     }
 }
